+++ app/assets/javascripts/gentelella/scrollview.js
@@ -1,0 +1,9 @@
+// #<{(|* ******  scrollview  *********************** *|)}>#
+$(document).ready(function () {
+  $(document).on('turbolinks:load', function() {
+    $(".scroll-view").niceScroll({
+        cursorcolor: "rgba(42, 63, 84, 0.35)"
+    });
+  });
+});
+// #<{(|* ******  /scrollview  *********************** *|)}>#