+++ app/jobs/rimrock/trigger_update_job.rb
@@ -4,7 +4,7 @@ module Rimrock
     queue_as :computation
 
     def perform
-      User.with_active_computations.each do |user|
+      User.with_submitted_computations.each do |user|
         Rimrock::UpdateJob.perform_later(user)
       end
     end