+++ config/initializers/plgrid_openid.rb
@@ -20,8 +20,8 @@ Devise.setup do |config|
                     AX[:POSTresponse]
                   ]
 
-  OpenID.fetcher.ca_file = File.join(Rails.root, 'config', 'ssl',
-                                     'DigiCertAssuredIDRootCA.pem')
+  OpenID.fetcher.ca_file = Rails.root.join('config', 'ssl',
+                                           'DigiCertAssuredIDRootCA.pem').to_s
 end
 
 module OpenID