+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 
 ### Changed
+- Update rubocop into 0.47.1, fix new discovered offenses (@mkasztelnik)
 
 ### Deprecated